% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_boost_tree_catboost.R
\name{add_boost_tree_catboost}
\alias{add_boost_tree_catboost}
\title{Wrapper to add \code{catboost} engine to the parsnip \code{boost_tree} model}
\usage{
add_boost_tree_catboost()
}
\description{
Catboost 패키지(엔진)를 boost_tree 모델에 부착한다.
}
