% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_catboost.R
\name{install_catboost}
\alias{install_catboost}
\title{Installation of Catboost Packages}
\usage{
install_catboost(version = "1.1.1")
}
\arguments{
\item{version}{catboost version that you want to install.}
}
\description{
Install \code{Catboost} Package By URL.
}
