% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_boost_tree_catboost.R
\name{add_boost_tree_catboost}
\alias{add_boost_tree_catboost}
\title{Wrapper to add \code{catboost} engine to the \code{parsnip} format}
\usage{
add_boost_tree_catboost()
}
\description{
Attach the \code{catboost} package to \code{boost_tree()} Function.
}
